+++ src/components/sales/status/tab/Tab.tsx
@@ -1,66 +1,21 @@
-import { getMonthsSales, getTodaySales, getWeekSales } from '@/server/api/supabase/sales';
-import { formatData } from '@/shared/helper';
+import { useCalendar } from '@/hooks/sales/useCalendar';
 import useSalesStore from '@/shared/store/sales';
-import moment from 'moment';
 import Calendar from './calendar/Calendar';
 import styles from './styles/tab.module.css';
+
 const Tab = () => {
   const {
-    date: { currentDate, utcStandardDate },
+    date: { currentDate },
     isShow,
-    setData,
-    setIsShow,
-    setCurrentDate,
-    setSelectedDate,
-    setRecord,
   } = useSalesStore();
 
-  const clickShowCalendarHandler = () => setIsShow(true);
-
-  const today = moment();
-  const yesterDay = today.clone().subtract(1, 'day');
-
-  const clickMoveYesterdayHandler = () => {
-    if (yesterDay === currentDate) return;
-    setCurrentDate(yesterDay);
-    setSelectedDate(yesterDay);
-  };
-
-  const clickMoveTodayHandler = async () => {
-    if (today === currentDate) return;
-    const { sales, formatType } = await getTodaySales(utcStandardDate.clone());
-    if (sales.length !== 0) {
-      const { result, recordData } = formatData(sales, formatType);
-      if (result && recordData) {
-        setData(result);
-        setRecord(recordData);
-      }
-    }
-    setCurrentDate(today);
-    setSelectedDate(today);
-  };
-
-  const clickWeeksChartHandler = async () => {
-    const { sales, formatType } = await getWeekSales(utcStandardDate.clone());
-    if (sales.length !== 0) {
-      const { result, recordData } = formatData(sales, formatType);
-      if (result && recordData) {
-        setData(result);
-        setRecord(recordData);
-      }
-    }
-  };
-
-  const clickMonthsChartHandler = async () => {
-    const { sales, formatType } = await getMonthsSales(utcStandardDate.clone());
-    if (sales.length !== 0) {
-      const { result, recordData } = formatData(sales, formatType);
-      if (result && recordData) {
-        setData(result);
-        setRecord(recordData);
-      }
-    }
-  };
+  const {
+    clickMoveYesterdayHandler,
+    clickMoveTodayHandler,
+    clickWeeksChartHandler,
+    clickMonthsChartHandler,
+    clickShowCalendarHandler,
+  } = useCalendar();
 
   return (
     <div className={styles['wrapper']}>
